refactor: view 패키지를 system 하위로 이동
