+++ src/main/java/subway/system/SectionSystem.java
@@ -3,7 +3,7 @@ package subway.system;
 import java.util.Scanner;
 import subway.manager.SectionManager;
 import subway.system.helper.SectionSystemInputValidator;
-import subway.view.SectionManagerView;
+import subway.system.view.SectionManagerView;
 
 public class SectionSystem {
