+++ src/main/java/subway/system/StationSystem.java
@@ -3,7 +3,7 @@ package subway.system;
 import java.util.Scanner;
 import subway.manager.StationManager;
 import subway.system.helper.StationSystemInputValidator;
-import subway.view.StationManagerView;
+import subway.system.view.StationManagerView;
 
 public class StationSystem {
